Spacemice support refactored to use a common base class
