+++ src/Gui/3Dconnexion/GuiNativeEventLinux.h
@@ -23,8 +23,7 @@
 #ifndef GUINATIVEEVENT_H
 #define GUINATIVEEVENT_H
 
-#include <vector>
-#include <QObject>
+#include "GuiAbstractNativeEvent.h"
 
 class QMainWindow;
 
@@ -32,10 +31,18 @@ namespace Gui
 {
 	class GUIApplicationNativeEventAware;
 
-	class GuiNativeEvent : public QObject
+	class GuiNativeEvent : public GuiAbstractNativeEvent
 	{
-#include "GuiNativeEventCommon.h"
-	private slots:
+	Q_OBJECT
+	public:
+		GuiNativeEvent(GUIApplicationNativeEventAware *app);
+		~GuiNativeEvent() override final;
+		void initSpaceball(QMainWindow *window) override final;
+	private:
+		GuiNativeEvent();
+		GuiNativeEvent(const GuiNativeEvent&);
+		GuiNativeEvent& operator=(const GuiNativeEvent&);
+	private Q_SLOTS:
 		void pollSpacenav();
 	};
 }
