+++ src/Gui/3Dconnexion/GuiNativeEventMac.h
@@ -23,11 +23,9 @@
 #ifndef GUINATIVEEVENT_H
 #define GUINATIVEEVENT_H
 
-#include <vector>
-#include <QObject>
+#include "GuiAbstractNativeEvent.h"
 
 class QMainWindow;
-class GUIApplicationNativeEventAware;
 
 #include <IOKit/IOKitLib.h>
 #include <ConnexionClientAPI.h>
@@ -46,9 +44,17 @@ namespace Gui
 {
 	class GUIApplicationNativeEventAware;
 
-	class GuiNativeEvent : public QObject
+	class GuiNativeEvent : public GuiAbstractNativeEvent
 	{
-	#include "GuiNativeEventCommon.h"
+	Q_OBJECT
+	public:
+		GuiNativeEvent(GUIApplicationNativeEventAware *app);
+		~GuiNativeEvent() override final;
+		void initSpaceball(QMainWindow *window) override final;
+	private:
+		GuiNativeEvent();
+		GuiNativeEvent(const GuiNativeEvent&);
+		GuiNativeEvent& operator=(const GuiNativeEvent&);
 	private:
         static UInt16 tdxClientID; /* ID assigned by the driver */
         static uint32_t lastButtons;
