+++ src/Gui/3Dconnexion/GuiNativeEventWin32.cpp
@@ -49,7 +49,6 @@ http://www.3dconnexion.com/forum/viewtopic.php?f=19&t=4968&sid=72c018bdcf0e6edc9
   #include "GuiRawInputEventFilter.h"
 #endif // #if QT_VERSION >= 0x050000
 
-std::vector<int> Gui::GuiNativeEvent::motionDataArray(6,0);
 Gui::GuiNativeEvent* Gui::GuiNativeEvent::gMouseInput = 0;
 
 
@@ -180,8 +179,8 @@ static const struct tag_VirtualKeys _3dmouseVirtualKeys[]=
 };
 
 Gui::GuiNativeEvent::GuiNativeEvent(Gui::GUIApplicationNativeEventAware *app)
+: GuiAbstractNativeEvent(app)
 {
-	mainApp = app;
 }
 
 Gui::GuiNativeEvent::~GuiNativeEvent()
